refactor shit
Сорри, не придумал нормального названия :)
